# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Arc Network Token Swap DApp API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /swap/estimate:
    post:
      operationId: estimateSwap
      tags: [swap]
      summary: Estimate swap rate
      description: Get a pre-swap estimate with price impact and fee breakdown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapRequest"
      responses:
        "200":
          description: Swap estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapEstimate"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /swap/execute:
    post:
      operationId: executeSwap
      tags: [swap]
      summary: Execute a token swap (backend peer-to-peer, fallback for cirBTC)
      description: Execute the swap on-chain after user approval. Used for cirBTC pairs and when the AMM pool is not available.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapExecuteRequest"
      responses:
        "200":
          description: Swap executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /swap/record:
    post:
      operationId: recordSwap
      tags: [swap]
      summary: Record a completed on-chain router swap
      description: Records a swap executed by the user directly through the Uniswap V2 router
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapRecordRequest"
      responses:
        "200":
          description: Swap recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SwapRequest:
      type: object
      required: [tokenIn, tokenOut, amountIn]
      properties:
        tokenIn:
          type: string
          description: Input token symbol (USDC, EURC, cirBTC)
        tokenOut:
          type: string
          description: Output token symbol (USDC, EURC, cirBTC)
        amountIn:
          type: string
          description: Amount to swap as a decimal string
    SwapExecuteRequest:
      type: object
      required: [tokenIn, tokenOut, amountIn, userAddress]
      properties:
        tokenIn:
          type: string
        tokenOut:
          type: string
        amountIn:
          type: string
        userAddress:
          type: string
          description: User's wallet address (0x...)
    SwapRecordRequest:
      type: object
      required: [tokenIn, tokenOut, amountIn, amountOut, userAddress, txHash]
      properties:
        tokenIn:
          type: string
        tokenOut:
          type: string
        amountIn:
          type: string
        amountOut:
          type: string
        userAddress:
          type: string
        txHash:
          type: string
    SwapEstimate:
      type: object
      required: [tokenIn, tokenOut, amountIn, estimatedAmountOut, priceImpact, platformFee, effectiveAmountIn, rate]
      properties:
        tokenIn:
          type: string
        tokenOut:
          type: string
        amountIn:
          type: string
        estimatedAmountOut:
          type: string
        priceImpact:
          type: string
          description: Price impact as a percentage string
        platformFee:
          type: string
          description: Platform fee amount
        effectiveAmountIn:
          type: string
          description: Amount after fee deduction
        rate:
          type: string
          description: Exchange rate
    SwapResult:
      type: object
      required: [success, transactionHash, explorerUrl, amountOut, fee]
      properties:
        success:
          type: boolean
        transactionHash:
          type: string
        explorerUrl:
          type: string
        amountOut:
          type: string
        fee:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]