# lib/api-zod/src/generated/types/swapRecordRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arc Network Token Swap DApp API
 * OpenAPI spec version: 0.1.0
 */

export interface SwapRecordRequest {
  tokenIn: string;
  tokenOut: string;
  amountIn: string;
  amountOut: string;
  userAddress: string;
  txHash: string;
}